requirements.txt
boto3==1.26.3
clearml==1.7.2
hydra-core==1.2.0
numpy==1.23.4
onnxruntime==1.13.1
pytorch-lightning==1.7.7
git+https://github.com/unball/python_simulator@master